requirements.txt
Flask==2.2.3
Pillow==7.2.0
opencv-python==4.7.0.72